TWAI-Backend/requirements.txt
# Core AI/ML
openai==1.97.0
langchain==0.3.26
langchain-openai==0.3.28
langchain-google-genai==2.1.8
langchain-pinecone==0.2.9
langchain-text-splitters==0.3.8
langchain-community==0.3.27
langchain-core==0.3.69
pinecone==7.3.0
cohere==5.16.1
google-ai-generativelanguage==0.6.18

# PDF Processing
llama-parse==0.6.50
llama-cloud==0.1.34
nest-asyncio==1.6.0
PyMuPDF==1.26.3
pypdf==5.8.0

# Web Framework
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20

# AWS
boto3==1.39.9
botocore==1.39.9

# Data Processing
numpy>=2.1.0
pandas==2.2.3
matplotlib==3.10.3

# NLP
spacy==3.8.7
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85

# HTTP and Utilities
requests==2.32.4
python-dotenv==1.1.1

# Essentials
pydantic==2.11.7
typing_extensions==4.14.1
certifi==2025.7.14
